refactoring name() to title(), subblocks are better formatted, html output, daily activity visualization is back, regression: crash at malformed card file(uchnast)
git-svn-id: svn://svn.code.sf.net/p/readesm/code@317 7258aa16-79bd-4fdc-8f83-50ad731c2ab6

+++ fileformat/PictureGenerators/SvgDayActivity.h
@@ -7,7 +7,7 @@ class SvgDayActivity : public SvgDayGraph {
 	Q_DECLARE_TR_FUNCTIONS(SvgDayActivity)
 public:
 	///add an activity rectangle. from is minutes since midnight, duration is minutes, height pixels, color html code, and title text to show on mouseover
-	virtual void add(int from, int duration, int height, const QString& color, const QString& title);
+	virtual void add(int from, int duration, float height, const QString& color, const QString& title);
 	virtual QString toString() const;
 protected:
 	static QString drawLegend();
